requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Js2Py==0.74
MarkupSafe==2.1.2
packaging==23.1
Pillow==9.5.0
pyjsparser==2.7.1
pytesseract==0.3.10
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.29.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==3.14.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
